Ejercicio 9 terminado

+++ ejercicios-pyqt5/ejercicio_nueve.py
@@ -12,17 +12,29 @@ class MiVentana(QMainWindow):
         self.btnAgregar.clicked.connect(self.agregar_persona)
         self.btnEditar.clicked.connect(self.editar_persona)
         self.btnEliminar.clicked.connect(self.eliminar_persona)
-
+#!-----------------------------------------------------------------------------------------------------
     def agregar_persona(self):
-        self.tablaDatos.insertRow(0)
-        self.tablaDatos.setItem(0, 0, QTableWidgetItem(self.nombre.text()))
-        self.tablaDatos.setItem(0, 1, QTableWidgetItem(self.apellido.text()))
-        self.tablaDatos.setItem(0, 2, QTableWidgetItem(self.email.text()))
-        self.borrar_entradas() 
-
+            if len(self.nombre.text()) == 0:
+                msg = QMessageBox(QMessageBox.Warning, "Error", "Tienes que completar el nombre")
+                msg.exec()
+            if len(self.apellido.text()) == 0:
+                msg = QMessageBox(QMessageBox.Warning, "Error", "Tienes que completar el apellido")
+                msg.exec()
+            if len(self.email.text()) == 0:
+                msg = QMessageBox(QMessageBox.Warning, "Error", "Tienes que completar el email")
+                msg.exec()
+            if len(self.nombre.text()) and len(self.apellido.text()) and len(self.email.text()) != 0:
+                self.tablaDatos.insertRow(0)
+                self.tablaDatos.setItem(0, 0, QTableWidgetItem(self.nombre.text()))
+                self.tablaDatos.setItem(0, 1, QTableWidgetItem(self.apellido.text()))
+                self.tablaDatos.setItem(0, 2, QTableWidgetItem(self.email.text()))
+                self.borrar_entradas()
+#!-----------------------------------------------------------------------------------------------------
     def editar_persona(self):
         fila = self.tablaDatos.currentRow()
         colum = self.tablaDatos.currentColumn()
+
+        #? Nombre
         if colum == 0:
             texto, ok = QInputDialog.getText(self, "Editar", "Nuevo nombre")
             if ok and texto != '':
@@ -30,6 +42,7 @@ class MiVentana(QMainWindow):
             else:
                 msg = QMessageBox(QMessageBox.Warning, "Error", "Tienes que completar el nombre")
                 msg.exec()
+        #? Apellido
         if colum == 1:
             texto, ok = QInputDialog.getText(self, "Editar", "Nuevo apellido")
             if ok and texto != '':
@@ -37,6 +50,7 @@ class MiVentana(QMainWindow):
             else:
                 msg = QMessageBox(QMessageBox.Warning, "Error", "Tienes que completar el nombre")
                 msg.exec()
+        #? Email
         if colum == 2:
             texto, ok = QInputDialog.getText(self, "Editar", "Nuevo email")
             if ok and texto != '':
@@ -44,14 +58,16 @@ class MiVentana(QMainWindow):
             else:
                 msg = QMessageBox(QMessageBox.Warning, "Error", "Tienes que completar el nombre")
                 msg.exec()
-
-
-        #self.tablaDatos.item()
-
+#!-----------------------------------------------------------------------------------------------------
     def eliminar_persona(self):
-        fila = self.tablaDatos.currentRow()
-        self.tablaDatos.removeRow(fila)
-
+        msg = QMessageBox(QMessageBox.Warning, "Eliminar", "Seguro quieres eliminarlo?", QMessageBox.Yes | QMessageBox.No)
+        resultado = msg.exec()
+        if resultado == QMessageBox.Yes: 
+            fila = self.tablaDatos.currentRow()
+            self.tablaDatos.removeRow(fila)
+        if resultado == QMessageBox.No:
+            pass
+#!-----------------------------------------------------------------------------------------------------
     def borrar_entradas(self):
         self.nombre.clear()
         self.apellido.clear()
